salvage/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfdmini"
version = "0.0.1"
description = "Minimal evolving CFD solver sandbox"
authors = [{name="Dev"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "numpy>=1.26,<3.0",
  "pyyaml>=6.0",
  "nbformat>=5.9",
  "pytest>=8.0"
]

[project.optional-dependencies]
dev = [
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","W","I","B","UP"]
ignore = ["E203","E266","E501"]

[tool.ruff.isort]
known-first-party = ["framework"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
exclude = ["tests/"]

[tool.pytest.ini_options]
addopts = "-q"
